#! Refactoring and code cleanup
