+++ router/ActionTypes.js
@@ -1,0 +1,31 @@
+/**
+ * Name of actions that can trigger routing
+ *
+ * @enum {string}
+ * @type {Object<string, string>}
+ */
+export const ActionTypes = Object.freeze({
+  /**
+   * @const
+   * @type {string}
+   */
+  REDIRECT: 'redirect',
+
+  /**
+   * @const
+   * @type {string}
+   */
+  CLICK: 'click',
+
+  /**
+   * @const
+   * @type {string}
+   */
+  POP_STATE: 'popstate',
+
+  /**
+   * @const
+   * @type {string}
+   */
+  ERROR: 'error'
+});
