+++ router/ClientRouter.js
@@ -1,6 +1,7 @@
 // @client-side
 
 import AbstractRouter from './AbstractRouter';
+import { ActionTypes } from './ActionTypes';
 import RouteFactory from './RouteFactory';
 import Dispatcher from '../event/Dispatcher';
 import PageManager from '../page/manager/PageManager';
@@ -29,38 +30,6 @@ const Events = Object.freeze({
    * @type {string}
    */
   POP_STATE: 'popstate'
-});
-
-/**
- * Name of actions that can trigger routing
- *
- * @enum {string}
- * @type {Object<string, string>}
- */
-export const ActionTypes = Object.freeze({
-  /**
-   * @const
-   * @type {string}
-   */
-  REDIRECT: 'redirect',
-
-  /**
-   * @const
-   * @type {string}
-   */
-  CLICK: 'click',
-
-  /**
-   * @const
-   * @type {string}
-   */
-  POP_STATE: 'popstate',
-
-  /**
-   * @const
-   * @type {string}
-   */
-  ERROR: 'error'
 });
 
 /**
