+++ router/__tests__/ClientRouterSpec.js
@@ -1,6 +1,7 @@
 import Dispatcher from 'event/Dispatcher';
 import PageManager from 'page/manager/PageManager';
-import ClientRouter, { ActionTypes } from 'router/ClientRouter';
+import { ActionTypes } from 'router/ActionTypes';
+import ClientRouter from 'router/ClientRouter';
 import RouteFactory from 'router/RouteFactory';
 import Window from 'window/Window';
